Fix slpFolderStream
