+++ src/stream/index.ts
@@ -1,3 +1,3 @@
 export * from "./slpFolderStream";
 export * from "./slpLiveStream";
-export * from "./slpStream";
+export * from "./rxSlpStream";
